feat(config): continously update config from remote URL
These changes makes it possible to configure a remote URL to provide a
new config. In practise that allows for capture, camera and other options
to be changed remotely over time.

That's valuable when the raspberry device is not directly exposed on the
internet, therefore SSHing in to change the config is not an option.