fix(deps): set go module name to meinside/rpi-camera-timelapse-go
Primarily to ensure that references to internal subpackages of this project
is resolved locally, rather than using the upstream / forked project as
internal subpackage.

To use `phillipj/rpi-camera-timelapse-go` as module name, all the internal
subpackage imports has to be renamed as well -- probably worth considering
soonish as I'm not making an effort in following the upstream project.